Set up source files so they can be easily deployed on dokploy
Remove caddy, set up compose.prod.yml for dokploy.
Remove other confusing configs and unneeded files, keep the configs
minimal and with dokploy in mind.

Change the main Makefile to be simpler and reference the most used
commands needed.

Add some more env vars for the AI models and the scrapping/input news
generation so that we can control from the deployment and not only by
code, also for the cron itself.

Fix problem with API calls (e.g. popular articles on main page) not
getting called cuz of wrong building of the api url args.