Next build complained about account deletion render
tested that the new code renders and the app can be built.
also needed to fix the gdpr and cookie-policy pages for fetchtopics not
available at the static build of next, similar to about page situation.
for now only the static stuff is got, will deal with it more properly
later.

also fix the link in the delete email to have the correct api endpoint url.